Vérification des types pour : Pour, Allantde, Switch

+++ LACompiler/LACompiler/Structures.h
@@ -167,6 +167,6 @@ inline string enMiniscule(string chaine) {
 typedef struct Identificateur {
 	string nom;
 	string type;
-	int taille1;
-	int taille2;
+	long taille1;
+	long taille2;
 }identificateur;
